refactor: replace detailed run page with dual-panel layout for website analysis
- Replaces the extensive inline run details page with a simplified dual-panel component
- Implements a new page that uses context providers and components for UI and debug info
- Adds a new dual-panel page component with layout, analysis UI, and run details
- Redirects from the old detailed page to the new dual-panel view for better UX
- Removes legacy code related to real-time updates, loading states, and detailed rendering
- Enhances maintainability by modularizing the UI into shared components and layout
- Updates routing to streamline navigation to the dual-panel interface
