+++ examples/playground/app/websites/runs/[run_id]/dual-panel/page.tsx
@@ -1,0 +1,51 @@
+'use client';
+
+import { useParams } from 'next/navigation';
+import { FlowRunProvider, useFlowRun } from '@/components/flow-run-provider';
+import FlowRunDetails from '@/components/flow-run-details';
+import WebsiteAnalysisUI from '@/components/website-analysis-ui';
+
+// Component that uses the shared context
+function DualPanelContent() {
+  const { runData, loading, error, currentTime, analyzeWebsite } = useFlowRun();
+
+  return (
+    <div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
+      {/* Left panel: User-friendly UI */}
+      <div className="lg:col-span-1">
+        <WebsiteAnalysisUI
+          runData={runData}
+          loading={loading}
+          error={error}
+          onAnalyzeWebsite={analyzeWebsite}
+        />
+      </div>
+
+      {/* Right panel: Technical debug UI */}
+      <div className="lg:col-span-2">
+        <FlowRunDetails
+          runId={runData?.run_id || ''}
+          runData={runData}
+          loading={loading}
+          error={error}
+          currentTime={currentTime}
+        />
+      </div>
+    </div>
+  );
+}
+
+export default function DualPanelPage() {
+  const params = useParams();
+  const runId = params.run_id as string;
+
+  return (
+    <div className="container mx-auto py-8">
+      <h1 className="text-3xl font-bold mb-6">Website Analysis</h1>
+      
+      <FlowRunProvider runId={runId}>
+        <DualPanelContent />
+      </FlowRunProvider>
+    </div>
+  );
+}
