+++ examples/playground/package.json
@@ -19,6 +19,7 @@
     "autoprefixer": "10.4.20",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "framer-motion": "^12.9.2",
     "lucide-react": "^0.468.0",
     "next": "latest",
     "next-themes": "^0.4.3",
